ai_service/requirements.txt
flask==2.3.2
opencv-python==4.7.1.72
ultralytics==8.0.120
transformers==4.30.2
torch==2.0.1
torchvision==0.15.2
pillow==9.5.0
numpy==1.24.3
requests==2.31.0
python-dotenv==1.0.0